Initial work to get JobManagement class to accept...
- a job name and index (work on that one instance)
- a job name (work on all instances of a job)
- no job name (work on all instances of all jobs)

Currently failing spec is just around messaging for the "all jobs" case.

Next up would be to be sure the call to Bosh::Cli::Client::Director
makes sense, and have it construct a request.  Then, of course, to
handle that request on the director API.

(Actually, it makes better sense to build in the API support first, then
add the CLI capability...)

[#100244046](https://www.pivotaltracker.com/story/show/100244046)